web_opt/pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-web_opt"
version = "0.1.0"
description = "WebOpt environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv dependencies (required for server functionality)
    # "openenv-core @ git+https://github.com/meta-pytorch/OpenEnv.git@main#subdirectory=src/core",
    "openenv-core>=0.1.0",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    "mcp",
    "Pillow",
    "scikit-image",
    # Environment-specific dependencies
    # Add all dependencies needed for your environment here
    # Examples:
    # "numpy>=1.19.0",
    # "torch>=2.0.0",
    # "gymnasium>=0.29.0",
    # "openspiel>=1.0.0",
    # "smolagents>=1.22.0,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m web_opt.server.app
server = "web_opt.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["web_opt", "web_opt.server"]
package-dir = { "web_opt" = ".", "web_opt.server" = "server" }